Added cached sps functions Added generate Ascii functions

+++ scripts/functions/SpssFunctions.ps1
@@ -1,5 +1,6 @@
 ﻿. "$PSScriptRoot\LoggingFunctions.ps1"
 . "$PSScriptRoot\FileFunctions.ps1"
+. "$PSScriptRoot\CloudFunctions.ps1"
 
 function AddSpssFilesToDeliveryPackage {
     param(
@@ -21,35 +22,40 @@ function AddSpssFilesToDeliveryPackage {
         throw "No instrument name argument provided"
     }
 
-    # Copy Manipula spss files to the processing folder
-    Copy-Item -Path "$PSScriptRoot\..\manipula\spss\*" -Destination $processingFolder
+    if (-not (Test-Path "$processingFolder\spss"))
+    {
+        # Copy Manipula spss files to the processing folder
+        Copy-Item -Path "$PSScriptRoot\..\manipula\spss\*" -Destination $processingFolder
 
-    # Generate SPS file
-    try {
-        & cmd.exe /c $processingFolder\Manipula.exe "$processingFolder\GenerateStatisticalScript.msux" -K:meta="$instrumentName.bmix" -H:"" -L:"" -N:oScript="$instrumentName,iFNames=,iData=$instrumentName.bdix" -P:"SPSS;;;;;;$instrumentName.asc;;;2;;64;;Y" -Q:True
-        LogInfo("Generated the .SPS file")
+        # Generate SPS file
+        try {
+            & cmd.exe /c $processingFolder\Manipula.exe "$processingFolder\GenerateStatisticalScript.msux" -K:meta="$instrumentName.bmix" -H:"" -L:"" -N:oScript="$instrumentName,iFNames=,iData=$instrumentName.bdix" -P:"SPSS;;;;;;$instrumentName.asc;;;2;;64;;Y" -Q:True
+            LogInfo("Generated the .SPSS file")
+            #create an sps folder
+            CreateANewFolder -folderPath $processingFolder -folderName "spss"
+            #copying the files needed to create an Ascii file
+            Copy-Item -Path "$processingFolder\*.sps" -Destination "$processingFolder/spss"
+            #adding the above files to the delivery zip
+            AddFolderToZip -pathTo7zip $env:TempPath -folder "$processingFolder/spss" -zipFilePath $deliveryZip 
+            #uploading the delivery file back to the dqs bucket with the 
+            UploadFileToBucket -filePath $deliveryZip -bucketName "$($env:ENV_BLAISE_DQS_BUCKET)/$($instrumentName).bpkg"
+        }
+        catch {
+            LogWarning("Generating SPS and FPS Failed for $instrumentName : $($_.Exception.Message)")
+        }
     }
-    catch {
-        LogWarning("Generating SPS and FPS Failed for $instrumentName : $($_.Exception.Message)")
-    }
-
-    # Generate .ASC file
-    try {
-        & cmd.exe /c $processingFolder\Manipula.exe "$processingFolder\ExportData_$instrumentName.msux" -A:True -Q:True -O
-        LogInfo("Generated the .ASC file")
-    }
-    catch {
-        LogWarning("Generating ASCII Failed for $instrumentName : $($_.Exception.Message)")
+    else {
+        Copy-Item -Path "$processingFolder/spss/*" -Destination $processingFolder
     }
 
     if ([string]::IsNullOrEmpty($subFolder))
     {      
         # Add the SPS, ASC & FPS files to the instrument package
-        AddFilesToZip -pathTo7zip $env:TempPath -files "$processingFolder\*.sps","$processingFolder\*.asc","$processingFolder\*.fps" -zipFilePath $deliveryZip
-        LogInfo("Added .SPS, .ASC, .Fps Files to $deliveryZip")
+        AddFilesToZip -pathTo7zip $env:TempPath -files "$processingFolder\*.sps" -zipFilePath $deliveryZip
+        LogInfo("Added .SPS Files to $deliveryZip")
     }
     else {
-        Copy-Item -Path "$processingFolder\*.sps","$processingFolder\*.asc","$processingFolder\*.fps" -Destination $subFolder
+        Copy-Item -Path "$processingFolder\*.sps" -Destination $subFolder
 
         AddFolderToZip -pathTo7zip $env:TempPath -folder $subFolder -zipFilePath $deliveryZip  
         LogInfo("Added '$subFolder' to '$deliveryZip'")
